pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arachnarium"
version = "1.0.0"
authors = [
  { name="Aleksei Stafeev", email="aleksei.stafeev@cispa.de" },
]
description = "A benchmarking platform for web scanners"
dependencies = [
  "PyYAML==6.0.*",
  "rich==13.7.*",
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/pixelindigo/arachnarium"
Issues = "https://github.com/pixelindigo/arachnarium/issues"

[project.scripts]
arachnarium = "arachnarium:main"
